rag_service/requirements.txt
langchain
langchain-community
langchain-classic
langchain-huggingface
langchain-groq
langchain-chroma
pypdf
python-dotenv
flask
flask-cors
sentence-transformers
